// server/src/enums/GameEvent.ts
/**
 * Enum representing every timed event in the game client.
 *
 * @enum {string}
 */
export enum GameEvent {
  WMI1 = "wmi1",
  WMI2 = "wmi2",
  MONSTER_MADNESS = "monsterMadness",
  BRUKKARG_WAR = "brukkargWar",
  CREATURE_CARNAGE = "creatureCarnage",
  MONSTER_BLITZKRIEG = "monsterBlitzkrieg",
  KING_OF_THE_HILL = "kingOfTheHill",
  HELL_RAISERS = "hellRaisers",
}

/**
 * Enum representing the two event slots in a calendar month.
 *
 * @enum {string}
 */
export enum EventSlot {
  A = "A",
  B = "B",
}